requirements.txt
appdirs==1.4.4
beautifulsoup4==4.11.2
bs4==0.0.1
certifi @ file:///C:/b/abs_85o_6fm0se/croot/certifi_1671487778835/work/certifi
charset-normalizer==3.0.1
distlib==0.3.0
filelock==3.0.12
idna==3.4
lxml==4.9.2
numpy==1.24.1
pandas==1.5.3
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
requests==2.28.2
six==1.16.0
soupsieve==2.3.2.post1
urllib3==1.26.14
virtualenv==20.0.23
wincertstore==0.2
